Initial commit of MemberMouse Groups Addon
With the groups addon you can define different types of groups allowing
customers to create their own group with a predefined number of seats
and use a signup link so members can join their group. This makes it so
that a single customer can pay for multiple seats and then members can
join their group for free or for a price based on how you configure the
group type. In order to use the groups extension plugin you must have
the MemberMouse plugin installed and activated.